Doom: Switched from C-, to M-, for localleader-alt
I wanted the leader and the leader to be consistent with their "alt"
version. So both should either use the control key or the meta key.
Unfortunately, C-SPC is already occupied by a bunch of completion calls
to enable completion in cases where completion does not trigger
automatically. In my case that means for example org files.