i3: Work layout uses Emacs Clients
Instead of the "standalone" Emacs window, the clients connect to the
running daemon of Emacs.

Always connecting to the daemon is consistent. However, it might be
desirable to have all the main workflows in one Emacs frame with their
respective workspaces instead of having two Emacs frames on i3
workspaces 1 and 3.