kmonad: Move to AltGr for better interoperability
Using AltGr makes it easy to set up umlauts with the moonlander
keyboard. At the same time, kmonad now replaces the physical right alt
key with the left alt key in order to retain the alt behavior under the
right thumb when working on the laptop directly.

It will be interesting to see how this works on a mac.